requirements.txt
openai
requests
SpeechRecognition
pydub
python-dotenv
gTTS
urllib3
Flask
flask_apscheduler
line-bot-sdk
deepface
beautifulsoup4